refactor: 쿠폰 모달에서 날짜 포맷팅 및 쿠폰 세부 정보 렌더링 로직 개선
