+++ README.md
@@ -37,7 +37,7 @@
 역할과 책임
 단일 책임 원칙
 
-- [ ]컴포넌트/훅을 잘게 나누고, 각 함수가 하나의 책임만 가지고 있는지 점검한다.
+- [x]컴포넌트/훅을 잘게 나누고, 각 함수가 하나의 책임만 가지고 있는지 점검한다.
 - [ ]컴포넌트는 UI 표현만 담당하며, 상태 관리 로직이나 파생 상태 계산 로직(e.g. 쿠폰 적용 금액 계산)을 포함하지 않아야 한다.
 - [ ]동일한 값에 대한 계산(e.g. 할인 금액, 배송비 등)은 항상 한 곳에서만 수행되도록 하여 같은, 중복된 책임이 발생하지 않도록 한다.
